requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
PyYAML==3.10
Werkzeug==0.9.4
docopt==0.6.1
itsdangerous==0.23
livereload==1.0.1
numpy==1.8.0
matplotlib==1.3.1
wsgiref==0.1.2
CairoSVG==1.0.3
tinycss==0.3
cssselect==0.9.1
loremipsum==1.0.2
Flask-WTF==0.8.3
Flask-DebugToolbar==0.8.0